+++ fe/fe-core/src/test/java/org/apache/doris/plugin/TestDialectPlugin1.java
@@ -19,28 +19,24 @@ package org.apache.doris.plugin;
 
 import org.apache.doris.analysis.StatementBase;
 import org.apache.doris.nereids.parser.Dialect;
+import org.apache.doris.nereids.parser.NereidsParser;
 import org.apache.doris.qe.SessionVariable;
 
 import com.google.common.collect.ImmutableSet;
 
 import java.util.List;
 
-public class TestHiveDialectPlugin extends Plugin implements DialectConverterPlugin {
+public class TestDialectPlugin1 extends Plugin implements DialectConverterPlugin {
 
-    private static final String TEST_CONVERTED_SQL = "select 1";
+    private static final String TEST_CONVERTED_SQL = "select 2";
 
     @Override
     public ImmutableSet<Dialect> acceptDialects() {
-        return ImmutableSet.of(Dialect.HIVE);
-    }
-
-    @Override
-    public String convertSql(String originSql, SessionVariable sessionVariable) {
-        return TEST_CONVERTED_SQL;
+        return ImmutableSet.of(Dialect.SPARK_SQL);
     }
 
     @Override
     public List<StatementBase> parseSqlWithDialect(String sql, SessionVariable sessionVariable) {
-        return null;
+        return new NereidsParser().parseSQL(TEST_CONVERTED_SQL, sessionVariable);
     }
 }
